requirements.txt
fastapi~=0.115.12
uvicorn
pandas~=2.2.3
openpyxl
instaloader~=4.14.1

requests~=2.32.3
beautifulsoup4~=4.13.3
python-dotenv
